// src/main/types/debugger.d.ts
export type LocationByUrl = {
    url: string;
    lineNumber: number;
    columnNumber: number;
};

export type MemoryValue = {
    rss: number;
    heapTotal: number;
    heapUsed: number;
    external: number;
};

export type CallFrames = {
    callFrameId: string;
    functionName: string;
    functionLocation: any;
    location: any;
    url: string;
    scopeChain: any[];
    this: any;
    canBeRestarted: true;
};

export type DebuggingResponse = {
    id?: number;
    method?: string;
    result?: {
        result: {
            type: "object";
            value: MemoryValue;
        };
    };
    params?: {
        callFrames?: CallFrames[];
        scriptId?: string;
        reason: string;
        url?: string;
        hitBreakpoint?: any[];
        sourceMapURL?: string;
    };
};

export type JsonDebuggerInfo = {
    description: string;
    devtoolsFrontendUrl: string;
    devtoolsFrontendUrlCompat: string;
    faviconUrl: string;
    id: string;
    title: string;
    type: string;
    url: string;
    webSocketDebuggerUrl: string;
};